# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AAS AI Backend
  description: AI-powered copilot API for automatic door technicians
  version: 1.0.0
  contact:
    name: Automatic Access Solutions
    url: https://aas-portal.com
servers:
  - url: https://api.aas-portal.com
    description: Production API
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/diagnose:
    post:
      operationId: diagnose
      summary: Quick diagnosis
      description: Non-streaming diagnostic endpoint optimized for speed
      tags: [Diagnose]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [door_id, symptom]
              properties:
                door_id:
                  type: string
                  description: Door registry Name (e.g., MH-1.81)
                symptom:
                  type: string
                  description: Description of the issue
                context:
                  $ref: '#/components/schemas/PageContext'
      responses:
        '200':
          description: Diagnosis response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 JWT access token
  schemas:
...
    PageContext:
      type: object
      properties:
        page:
          type: string
          description: Current page identifier
        door_id:
          type: string
          description: Door registry Name
        manufacturer:
          type: string
          enum: [Horton, Stanley, NABCO, Besam, Other]
        model:
          type: string
        door_type:
          type: string
          enum: [sliding, swing, folding, revolving, other]
        controller:
          type: string
        site_name:
          type: string
    ChatResponse:
      type: object
      required: [response_text]
      properties:
        response_text:
          type: string
        diagnosis:
          $ref: '#/components/schemas/DiagnosisBlock'
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        parts_needed:
          type: array
          items:
            $ref: '#/components/schemas/PartNeeded'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
    DiagnosisBlock:
      type: object
      required: [likely_cause, confidence, category]
      properties:
        likely_cause:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        category:
          type: string
    ChecklistItem:
      type: object
      required: [step, action]
      properties:
        step:
          type: integer
          minimum: 1
        action:
          type: string
        manual_ref:
          type: object
          properties:
            manual_id:
              type: string
            page:
              type: integer
    PartNeeded:
      type: object
      required: [part_number, description]
      properties:
        part_number:
          type: string
        description:
          type: string
        quantity:
          type: integer
          minimum: 1
          default: 1
    Source:
      type: object
      required: [type, relevance]
      properties:
        type:
          type: string
          enum: [playbook, manual]
        id:
          type: string
        manual_id:
          type: string
        page:
          type: integer
        relevance:
          type: number
          minimum: 0
          maximum: 1
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Authentication required
...